+++ foodblogkitchen-toolkit.php
@@ -4,7 +4,7 @@
  * Plugin Name:     Foodblog-Toolkit
  * Author URI:      https://www.foodblogr.com/
  * Description:     Toolkit for your Foodblog to optimize your blog for search engines. Including a Recipe block for the Gutenberg editor.
- * Version:         1.7.2
+ * Version:         1.8.0
  * Author:          foodblogr.com
  * Text Domain:     foodblogkitchen-toolkit
  * Domain Path:     /languages